network_test/requirements.txt
fastapi
uvicorn
requests
pyyaml
netmiko
